requirements.txt
annotated-types==0.7.0
anyio==4.9.0
browserbase==1.4.0
certifi==2025.7.14
distro==1.9.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
playwright==1.54.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
python-dotenv==1.0.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
